backend/pyproject.toml
[tool.poetry]
name = "optimart-django"
version = "0.1.0"
description = ""
authors = ["Isaac Adewumi", "Lijadu Obafunmilayo", "Olayinka Ganiyu"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-filter = "^23.2"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
huey = "^2.4.5"
django-cors-headers = "^4.2.0"
redis = "^5.0.0"
xrpl-py = "^2.2.0"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.284"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
skip-string-normalization = true
target-version = ['py311']
exclude = '''
  /(
    | \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
'''

[tool.ruff]
line-length = 119
select = [
    "E",
    "F",
    "W",
    "C",
    "N",
    "B",
    "G",
    "ISC",
    "ICN",
    "T",
    "UP",
    "YTT",
    "S",
    "BLE",
    "FBT",
    "A",
    "Q",
    "ERA",
    "PGH",
    "RET",
    "SIM",
    "TID",
    "ARG",
    "PL",
    "PIE",
    "PYI",
    "COM",
    "DTZ",
    "EXE",
    "INP",
    "PT",
    "TCH",
    "PTH",
    "INT",
    "TRY",
    "RSE",
    "SLF",
    "RUF",
]
exclude = ["migrations", ".venv"]
ignore = ["TRY003", "G004"]

[tool.ruff.per-file-ignores]
"**/test_**.py" = ["S101"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"
docstring-quotes = "double"
multiline-quotes = "double"

[tool.isort]
profile = "black"
line_length = 119
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
known_django = "django*"
combine_as_imports = true
lines_between_sections = 1
include_trailing_comma = true
known_rest_framework = "rest_framework"
skip_glob = ["**/.venv/*", "**/migrations/*"]
sections = ["FUTURE","STDLIB","THIRDPARTY","DJANGO", "REST_FRAMEWORK", "FIRSTPARTY","LOCALFOLDER"]
